agregando evaluaciones

+++ Enitdades/Evaluaciones.cs
@@ -12,7 +12,7 @@ namespace CoreEscuela.Enitdades
         public string Nombre { get; set; }
         public Alumno Alumno { get; set; }
         public Asignatura Asignatura { get; set; }
-        public double Nota { get; set; }
+        public float Nota { get; set; }
         public Evaluaciones()
         {
             UniqueId = Guid.NewGuid().ToString();
